backend/requirements.txt
# Project Vanguard Backend Dependencies

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic-settings>=2.4.0
python-dotenv>=1.0.0

# OpenTelemetry — Observability
opentelemetry-api>=1.27.0
opentelemetry-sdk>=1.27.0
opentelemetry-instrumentation-fastapi>=0.48b0
opentelemetry-instrumentation-httpx>=0.48b0
opentelemetry-instrumentation-logging>=0.48b0
opentelemetry-instrumentation-sqlalchemy>=0.48b0
opentelemetry-exporter-otlp-proto-grpc>=1.27.0
opentelemetry-propagator-b3>=1.27.0

# AI & RAG — LangChain ecosystem
openai>=1.40.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-pinecone>=0.1.0
langchain-text-splitters>=0.2.0

# Vector Store
pinecone-client>=4.1.0

# API Clients
httpx>=0.27.0
azure-storage-blob>=12.23.1

# Database & Auth
SQLAlchemy>=2.0.36
alembic>=1.14.0
asyncpg>=0.30.0
psycopg[binary]>=3.2.3
pwdlib[argon2]>=0.2.1
PyJWT>=2.10.1
email-validator>=2.2.0
aiosqlite>=0.20.0
greenlet>=3.1.1

# Ingestion & Parsing
beautifulsoup4>=4.12.0
python-multipart>=0.0.9
lxml>=5.0.0
pypdf>=5.4.0

# Scheduling (auto-sync polling)
apscheduler>=3.10.0

# Observability
loguru>=0.7.0

# Rate Limiting — Phase 7
slowapi>=0.1.9

# Azure Speech Services — Voice Pipeline TTS
azure-cognitiveservices-speech>=1.37.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
